requirements.txt
Flask==2.0.1
requests==2.26.0
python-dotenv==0.19.0
Flask-Assets==2.0
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Werkzeug<3.0.0
bcrypt==3.2.0 